Role Management and Permissions is completed
A new Permission Library is written to handle roles and permission of a user.

+++ application/views/role/index.php
@@ -25,8 +25,15 @@
 									<td><?php echo $role->name;?></td>
 									<td><?php echo $role->permissions;?></td>
 									<td>
-										<button class="btn btn-warning" onclick="edit_role(<?php echo $role->id;?>)"><i class="fa fa-edit"></i></button>
-										<button class="btn btn-danger" onclick="delete_role(<?php echo $role->id;?>)"><i class="fa fa-trash-o"></i></button>
+										<?php if (!in_array('role_view', $this->session->userdata('permissions'))) { ?>
+											<button class="btn btn-info" onclick="view_role(<?php echo $role->id;?>)"><i class="fa fa-file-text-o"></i></button>
+										<?php } ?>
+										<?php if (!in_array('role_edit', $this->session->userdata('permissions'))) { ?>
+											<button class="btn btn-warning" onclick="edit_role(<?php echo $role->id;?>)"><i class="fa fa-edit"></i></button>
+										<?php } ?>
+										<?php if (!in_array('role_delete', $this->session->userdata('permissions'))) { ?>
+											<button class="btn btn-danger" onclick="delete_role(<?php echo $role->id;?>)"><i class="fa fa-trash-o"></i></button>
+										<?php } ?>
 									</td>
 								</tr>
 							<?php }?>
@@ -58,7 +65,9 @@
 								<input name="name" placeholder="Role Title" class="form-control" type="text">
 							</div>
 						</div>
-						
+						<div class="row form-group">
+							 <div class="form-check permissions-area col col-md-12"></div>
+						</div>
 					</div>
 				</form>
 			</div>
